Conflicts:
	school-service/src/main/java/school/service/CourseService.java

+++ school-service/src/main/java/school/service/CourseService.java
@@ -8,7 +8,6 @@ import org.springframework.security.access.annotation.Secured;
 import school.model.Course;
 
 public interface CourseService {
-
     @Secured("ROLE_TEACHER")
-    List<Course> getCourseByUserIdAndDataRange(int id, Date from, Date till);
+    List<Course> getCourseByUserIdAndDataRange(long id, Date from, Date till);
 }
